backend/requirement.txt
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.0
typing_extensions==4.12.2
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Markdown==3.7
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
sqlparse==0.5.0
typing_extensions==4.12.2asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
Markdown==3.7
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.0
typing_extensions==4.12.2
uritemplate==4.1.1
